Add & refactor api service class for currency view
